Only backup databases that are less than 100MB to AFS
First order approximation, this skips databases that are smaller
than 100MB if they have large index files, but this prevents the
AFS backup volumes from becoming overful.

This version also fails to backup databases that have spaces in
their names, because of a bash for loop. Switching to \0 delimiters
will correct this issue.